feat: fixed tags edit page
